refactor(java): :recycle: refactored classes to different packages
