+++ core-java/src/main/java/com/github/wasiqb/boyka/actions/drivers/AlertActions.java
@@ -14,7 +14,7 @@
  * copies or substantial portions of the Software.
  */
 
-package com.github.wasiqb.boyka.actions;
+package com.github.wasiqb.boyka.actions.drivers;
 
 import static com.github.wasiqb.boyka.actions.CommonActions.getDriverAttribute;
 import static com.google.common.truth.Truth.assertThat;
