+++ core-java/src/test/java/com/github/wasiqb/boyka/testng/ui/saucedemo/actions/SauceDemoActions.java
@@ -1,11 +1,11 @@
 package com.github.wasiqb.boyka.testng.ui.saucedemo.actions;
 
-import static com.github.wasiqb.boyka.actions.ClickableActions.withMouse;
-import static com.github.wasiqb.boyka.actions.ElementActions.onElement;
-import static com.github.wasiqb.boyka.actions.FingerActions.withFinger;
-import static com.github.wasiqb.boyka.actions.NavigateActions.navigate;
-import static com.github.wasiqb.boyka.actions.TextBoxActions.onTextBox;
-import static com.github.wasiqb.boyka.actions.WindowActions.onWindow;
+import static com.github.wasiqb.boyka.actions.drivers.NavigateActions.navigate;
+import static com.github.wasiqb.boyka.actions.drivers.WindowActions.onWindow;
+import static com.github.wasiqb.boyka.actions.elements.ClickableActions.withMouse;
+import static com.github.wasiqb.boyka.actions.elements.ElementActions.onElement;
+import static com.github.wasiqb.boyka.actions.elements.FingerActions.withFinger;
+import static com.github.wasiqb.boyka.actions.elements.TextBoxActions.onTextBox;
 import static com.github.wasiqb.boyka.enums.PlatformType.WEB;
 import static com.github.wasiqb.boyka.enums.SwipeDirection.UP;
 import static com.github.wasiqb.boyka.sessions.ParallelSession.getSession;
