+++ core-java/src/test/java/com/github/wasiqb/boyka/testng/ui/theinternet/FramesTest.java
@@ -16,10 +16,10 @@
 
 package com.github.wasiqb.boyka.testng.ui.theinternet;
 
-import static com.github.wasiqb.boyka.actions.ClickableActions.withMouse;
-import static com.github.wasiqb.boyka.actions.ElementActions.onElement;
-import static com.github.wasiqb.boyka.actions.FrameActions.onFrame;
-import static com.github.wasiqb.boyka.actions.NavigateActions.navigate;
+import static com.github.wasiqb.boyka.actions.drivers.FrameActions.onFrame;
+import static com.github.wasiqb.boyka.actions.drivers.NavigateActions.navigate;
+import static com.github.wasiqb.boyka.actions.elements.ClickableActions.withMouse;
+import static com.github.wasiqb.boyka.actions.elements.ElementActions.onElement;
 import static com.github.wasiqb.boyka.manager.DriverManager.closeDriver;
 import static com.github.wasiqb.boyka.manager.DriverManager.createDriver;
 import static com.github.wasiqb.boyka.testng.ui.theinternet.pages.FramesPage.framesPage;
