+++ website/docs/framework-docs/guides/ui/page-action.md
@@ -8,13 +8,11 @@ Since we have a common page object for all the three platforms (i.e.: Web. Andro
 Let's check out the following action class example below for our `LoginPage` class we created:
 
 ```java
-import static com.github.wasiqb.boyka.actions.DriverActions.navigate;
-import static com.github.wasiqb.boyka.actions.KeyboardActions.enterText;
-import static com.github.wasiqb.boyka.actions.MouseActions.clickOn;
-import static com.github.wasiqb.boyka.actions.VerifyDriverActions.verifyBrowserTitle;
-import static com.github.wasiqb.boyka.actions.VerifyDriverActions.verifyBrowserUrl;
-import static com.github.wasiqb.boyka.actions.VerifyElementActions.verifyElementDisplayed;
-import static com.github.wasiqb.boyka.actions.VerifyElementActions.verifyElementEnabled;
+import static com.github.wasiqb.boyka.actions.drivers.NavigateActions.navigate;
+import static com.github.wasiqb.boyka.actions.drivers.WindowActions.onWindow;
+import static com.github.wasiqb.boyka.actions.elements.ElementActions.onElement;
+import static com.github.wasiqb.boyka.actions.elements.FingerActions.withFinger;
+import static com.github.wasiqb.boyka.actions.elements.TextBoxActions.onTextBox;
 import static com.github.wasiqb.boyka.enums.PlatformType.WEB;
 import static com.github.wasiqb.boyka.sessions.ParallelSession.getSession;
 import static com.github.wasiqb.boyka.testng.ui.saucedemo.pages.LoginPage.loginPage;
@@ -33,26 +31,31 @@ public class SauceDemoActions {
 
   public void verifyLogin (final String userName, final String password) {
     verifyNavigateToSite ();
-    enterText (loginPage ().getUsername (), userName);
-    enterText (loginPage ().getPassword (), password);
-    clickOn (loginPage ().getLoginButton ());
+    onTextBox (loginPage ().getUsername ()).enterText (userName);
+    onTextBox (loginPage ().getPassword ()).enterText (password);
+    withMouse (loginPage ().getLoginButton ()).click ();
     verifyLoggedIn ();
   }
 
   private void verifyNavigateToSite () {
     if (this.platformType == WEB) {
       navigate ().to (URL);
-      verifyBrowserUrl ().startsWith (URL);
+      navigate ().verifyUrl ()
+        .startsWith (URL);
     }
   }
 
   private void verifyLoggedIn () {
     if (this.platformType == WEB) {
-      verifyBrowserUrl ().isEqualTo (format ("{0}/inventory.html", URL));
-      verifyBrowserTitle ().isEqualTo ("Swag Labs");
+      navigate ().verifyUrl ()
+        .isEqualTo (format ("{0}/inventory.html", URL));
+      onWindow ().verifyTitle ()
+        .isEqualTo ("Swag Labs");
     }
-    verifyElementDisplayed (homePage ().getMenuButton ()).isTrue ();
-    verifyElementEnabled (homePage ().getMenuButton ()).isTrue ();
+    onElement (homePage ().getMenuButton ()).verifyIsDisplayed ()
+      .isTrue ();
+    onElement (homePage ().getMenuButton ()).verifyIsEnabled ()
+      .isTrue ();
   }
 }
 ```
